tracing: Fix X-Ray header values
X-Ray values are expected by the SDKs to be PascalCase otherwise they
are not parsed and dropped.

Also, we must ensure to _set_ rather than _add_ the X-Ray header.

Finally, when X-Ray validation fails (because Envoy was given a bad
trace-id for example) we shouldn't let the exception bubble up.

Signed-off-by: Marco Magdy <mmagdy@gmail.com>